Separate secret fetching from definitions
